model_package/requirements/requirements.txt
numpy== 1.22.3
pandas==1.4.2
scikit-learn==1.0.2
tensorflow-cpu==2.8.0
tensorflow-recommenders==0.6.0
pydantic==1.8.2
strictyaml==1.6.1
setuptools>=20.6.8
ruamel.yaml==0.16.12
joblib==1.1.0